# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api public de l'application",
        "description": "cette documentation permet de décrire l'api public de l'application bgrfacile sur l'ensemble des ressources disponibles",
        "contact": {
            "email": "benaja.bendo02@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "server api v2"
        }
    ],
    "paths": {
...
        "/locations/{id}": {
            "get": {
                "description": "locations paginer",
                "operationId": "eb2776423337fc4ed7d8949a03fb5028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json avec une clé data"
                    }
                }
            },
            "put": {
                "description": "locations paginer",
                "operationId": "f6c358b2abff2b69cd77a9aab24a326c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json avec une clé data"
                    }
                }
            },
            "delete": {
                "description": "locations delete",
                "operationId": "43c5d3da20086c376fdda3ba8ea0322f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json avec une clé data"
                    }
                }
            }
        }
    }
}